+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 1.  [#4479](https://github.com/influxdata/chronograf/pull/4479): Add validation to Alert Rule messages
 1.  [#4599](https://github.com/influxdata/chronograf/pull/4599): Fix search results updating race condition
 1.  [#4605](https://github.com/influxdata/chronograf/pull/4605): Fix vertical stuck vertical scroll in firefox
+1.  [#4612](https://github.com/influxdata/chronograf/pull/4612): Fix issue with disappearing alias'
 
 ## v1.6.2 [2018-09-06]